To merge basic components
